[XLA:GPU] Find canonical send/recv ops for async events in IR emitter
Previously, the IR emitter relied on the channel ID to determine the async events key for partially pipelined async collectives.
The channel ID was used in a misleading way. Generally, it is supposed to be unique per op, which would cause breakage here.
As a result this use of channel ID caused many special cases in the verifier and also in the emitter.

PiperOrigin-RevId: 683797585